delete unnecessary utils file
