+++ Statsdash/aggregate_data/google.py
@@ -1,7 +1,6 @@
 from .base import AggregateData
 from Statsdash.analytics import GoogleAnalytics, third_party_metrics
-from Statsdash.utils import utils
-from Statsdash import config
+from Statsdash import config, utils
 
 TABLES = config.GOOGLE['TABLES']
 Metrics = GoogleAnalytics.Metrics
@@ -34,7 +33,7 @@ class SummaryData(AnalyticsData):
         data = super()._format_data(data, site)
         return self._apply_averages(data)
 
-    # TODO I wonder if this should be happening here.
+    # TODO This could be handled in the template.
     def _apply_averages(self, data):
         """
         Get average for metrics which are averages.
@@ -46,7 +45,6 @@ class SummaryData(AnalyticsData):
             * `dict`
         """
         len_sites = len(self.sites)
-        # TODO metrics could be turned into classes. Each class could handle average.
         av = data.get(Metrics.pv_per_sessions[1], 0) / len_sites
         data[Metrics.pv_per_sessions[1]] = av
         av = data.get(Metrics.avg_session_time[1], 0) / len_sites / 60.0
@@ -95,7 +93,6 @@ class ArticleData(AnalyticsData):
     def _format_data(self, data, site):
         data = super()._format_data(data, site)
 
-        # TODO clean logic
         path = data['path']
         title = data['title']
 
@@ -214,95 +211,6 @@ class SocialData(AnalyticsData):
     aggregate_key = Dimensions.social_network
     match_key = Dimensions.social_network
     limit = 15
-
-
-# TODO social chart should be different form the data class. should be in different module
-# NOTE gets period for every day between start and finish
-
-#     def social_chart(self):
-#
-#         current = self.period.start_date
-#         end = self.period.end_date
-#         dates = []
-#         while current <= end:
-#             current_range = utils.StatsRange("day", current, current)
-#             dates.append(current_range)
-#             current = current + timedelta(days=1)
-#
-#         data = {}
-#         network_data = {}
-#         for count, date in enumerate(dates):
-#             social = []
-#             network_social = []
-#             metrics = "ga:pageviews,ga:users,ga:sessions"
-#             for site in config.TABLES.keys():
-#                 rows = [analytics.rollup_ids(self.site_ids[site], date.get_start(), date.get_end(), metrics=metrics, dimensions=None, filters="ga:socialNetwork!=(not set)", sort="-ga:users")]
-#                 if rows[0]:
-#                     rows = self._remove_ga_names(rows)
-#                     for row in rows:
-#                         row = utils.convert_to_floats(row, ["pageviews", "users", "sessions"])
-#                     if site in self.sites:
-#                         social.extend(rows)
-#                     network_social.extend(rows)
-#                 else:
-#                     logger.debug("No data for site " + site + " on " + date.get_start() + " - " + date.get_end())
-#
-#             aggregate = utils.aggregate_data(social, ["pageviews", "users", "sessions"])
-#             network_aggregate = utils.aggregate_data(network_social, ["pageviews", "users", "sessions"])
-#
-#             data[date.get_start()] = aggregate
-#             network_data[date.get_start()] = network_aggregate
-#
-#         x_labels = []
-#         graph_data = {"users": [], "pageviews": [], "network_pageviews": [], "network_users": []}
-#         for range in dates:
-#             x_labels.append(range.get_start())
-#             graph_data["users"].append(data[range.get_start()]["users"])
-#             graph_data["pageviews"].append(data[range.get_start()]["pageviews"])
-#             graph_data["network_users"].append(network_data[range.get_start()]["users"])
-#             graph_data["network_pageviews"].append(network_data[range.get_start()]["pageviews"])
-#
-#         chart = utils.chart("Social Data", x_labels, graph_data, "Day", "Number")
-#         return chart
-
-
-# TODO This should work like social chart. should be in different module
-#     def device_chart(self, data):
-#         chart_data = {}
-#         x_labels = []
-#         for count, row in enumerate(data):
-#             for device in row["data"]:
-#                 try:
-#                     chart_data[device["device_category"]].append(utils.percentage(device["users"], row["summary"]["users"]))
-#                 except KeyError:
-#                     chart_data[device["device_category"]] = [utils.percentage(device["users"], row["summary"]["users"])]
-#
-#             x_labels.append(row["month"])
-#
-#         chart = utils.chart("Device Chart", x_labels, chart_data, "Month", "Percentage of Users")
-#         return chart
-
-
-# TODO This is a composite table
-# def referring_sites_table(self, num_articles):
-#     # gets the data
-#     sources = self._get_by_source()[:5]
-#     referrals = []
-#     black_ex = '|'
-#     black_string = black_ex.join(config.SOURCE_BLACK_LIST)
-#     regex = re.compile(black_string)
-#     for row in sources:
-#         source = row["source"]
-#         match = regex.search(source)
-#         if match:
-#             continue
-#         else:
-#             filter = "ga:source==%s" % source
-#             article = self.referral_articles(filter, num_articles)
-#             row["source"] = source
-#             row["articles"] = article
-#             referrals.append(row)
-#     return referrals
 
 
 def get_site_ids():
